perf: single-byte field indices for common bytecode strings (Phase 3)
- compile_abilities.py: reorder strings by frequency, top 254 use u8 index
- vm.rs: BcReader.read_idx() reads u8, 0xFE escape for >= 254
- All key/string index reads updated: key(), skip_value, read_value_from_bc, collect_json_map
- Bytecode: 140KB -> 111KB (20.5% smaller)
- Top 254 strings cover 88.7% of all references

+++ engine/tests/test_modules/condition_evaluation_test.rs
@@ -1,6 +1,6 @@
 use crate::helpers::*;
 
-fn base_heart02(db: &rabuka_engine::card::CardDatabase, card_id: i16) -> u32 {
+fn base_heart02(db: &rabuka_engine::card::CardDatabase, card_id: i16) -> u8 {
     db.get_card(card_id)
         .and_then(|c| c.base_heart.as_ref())
         .map(|bh| {
